backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database - MongoDB Async Driver
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Configuration Management
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Email Validation
email-validator==2.1.0

# CORS Support (included in FastAPI)
